requirements.txt
chainlit==1.1.404
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
notebook==7.2.2
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
langgraph==0.2.23